refactor: rename result
