+++ contracts/cosmwasm-vm/cw-xcall/src/types/mod.rs
@@ -2,7 +2,7 @@ pub mod call_request;
 pub mod config;
 pub mod message;
 pub mod request;
-pub mod response;
+pub mod result;
 pub mod storage_keys;
 
 pub const LOG_PREFIX: &str = "[xcall_app]:";
@@ -13,4 +13,4 @@ use common::rlp::{Decodable, Encodable};
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{to_binary, Binary};
 use request::CSMessageRequest;
-use response::CSMessageResult;
+use result::CSMessageResult;
